iteration of text filter

+++ src/numberFilter.js
@@ -32,7 +32,7 @@ define([
         if (typeof value === 'number') {
             valueAsNumber = value;
         } else {
-            valueAsNumber = parseInt(value);
+            valueAsNumber = parseFloat(value);
         }
 
         switch (this.filterType) {
@@ -79,7 +79,7 @@ define([
             filterText = null;
         }
         if (filterText) {
-            this.filterNumber = parseInt(filterText);
+            this.filterNumber = parseFloat(filterText);
         } else {
             this.filterNumber = null;
         }
